requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-decouple
aiofiles
Pillow
opencv-python-headless
ffmpeg-python
pydantic
pydantic-settings
email-validator
httpx
pytest
pytest-asyncio
transformers
torch
torchvision
torchaudio
kokoro>=0.9.4
soundfile
syncsdk
boto3
numpy